Do not be willing to run migrate-tests.sh without .env.testing

+++ bin/migrate-tests.sh
@@ -4,4 +4,9 @@
 
 set -e
 
+if [[ ! -f ".env.testing" ]]; then
+  echo ".env.testing not found — stopping."
+  exit 1
+fi
+
 APP_ENV=testing bin/php artisan migrate:fresh
